ChessBotUser.user.js: turn node limit into a number
it makes it easier for the user to set a precise number than sliding
on a huge range.

I don't know what is the limit yet, or if there is any limit as
stockfish can easily get million nodes per second while I get amazing
results with lc0 with node limit of 10.